Add features to MessageList: scrolling inversion, and a "load more" button
The scrolling inversion is a simple boolean prop passed as isInverted to MessageList
it does a few things: changes the buttonbadge to point up, changes default
onDownButtonBadgeClick to scroll to top, changes when onDownButtonBadge appears, changes
where the loadMoreButton

The LoadMoreButton appears at the end of the feed of messages. It serves as a way to load
more messages (e.g. from a file or database when there are too many to display at once).
It appears when the loadMoreButton boolean prop is set to true. It has a callback prop,
onLoadMoreClick which activates a function when clicked (input is event). It also is customizable
with loadMoreCustomButton which replaces the default JSX for the component.

Both of these new features are disabled by default. There is an example, ./example-inverted/,
which shows these in action.